requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.0.3
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
bleach==3.3.0
blinker==1.4
boto3==1.17.73
botocore==1.20.73
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.2
cryptography==3.4.7
cycler==0.10.0
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-admin-interface==0.16.1
django-allauth==0.44.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-countries==7.1
django-crispy==1.8.1
django-crispy-forms==1.11.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.11.1
django-tailwind==2.0.1
djangorestframework==3.12.4
dnspython==2.1.0
docx2pdf==0.1.7
email-validator==1.1.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
GoogleNews==1.5.7
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
newsapi-python==0.2.6
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
poyo==0.5.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pygame==2.0.1
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-slugify==4.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytube3==9.6.4
pytz==2021.1
pywin32==227
pywinpty==0.5.7
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.6.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
social-auth-core==4.1.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
stripe==2.56.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
WTForms==2.3.3
youtube-dl==2021.3.25
